README.md
The gut microbiota is known to play a crucial role in maintaining intestinal homeostasis. Alterations in the microbial community composition, also known as dysbiosis, may put health status in risk and increase susceptibility to diseases. Inflammatory bowel disease (IBD) is a chronic disease with periods of active and inactive inflammation that constitutes to an important health problem worldwide. It is divided in two subtypes: Crohn’s disease (CD) and ulcerative colitis (UC) that present similar symptoms but different clinical manifestations. Current and most common diagnostics methods relay on endoscopic procedures, which are relatively invasive.

IBD has been widely associated with an alteration of the gut microbiome composition. Nevertheless, there is no clear consensus on the microbial pattern characteristic of the disorders.

We analysed a cohort of 356 fecal samples from non-IBD controls (n=112) and IBD patients (n=244) over time, applying a 16S rRNA gene sequencing approach to compare the fecal microbiome composition. Fecal and fungal loads were also obtained using real-time PCR amplification of the ITS2 region and the 16S rRNA gene, respectively. We combined the microbial data, demographic and standard laboratory data to discriminate CD and UC and to predict disease relapse implementing the random forest machine learning method.

Our results indicate that bacterial loads were significantly different between UC patients and their relatives, between CD patients in relapse and remission, and between UC patients in relapse and remission. In terms of fungal loads, significant differences were found between CD and UC patients, and between UC patients in relapse and remission. 

Dysbiosis was found to be significantly greater in patients with CD than with UC, as evidenced by a lower diversity. The microbiome of UC patients was more similar to that of healthy controls. Differences were found in fourteen genera among all the groups analyzed. Finally, microbial genera and microbial load data combined with standard laboratory and demographic data allowed to obtain a good performance in a random forest models, reaching an average AUC
of 0.89 in the different combinations made.

We demonstrated in this study that the use microbiome data combined with clinical and demographic data could be a useful and non-invasive tool for helping IBD diagnosis.